[ads] General code health

+++ components/brave_ads/core/internal/account/account_util_unittest.cc
@@ -7,6 +7,8 @@
 
 #include "brave/components/brave_ads/core/internal/account/account_util.h"
 
+#include <cstddef>
+
 #include "base/test/scoped_feature_list.h"
 #include "brave/components/brave_ads/core/internal/common/test/test_base.h"
 #include "brave/components/brave_ads/core/internal/settings/settings_test_util.h"
@@ -18,9 +20,9 @@ class BraveAdsAccountUtilTest : public test::TestBase {};
 
 TEST_F(BraveAdsAccountUtilTest, AlwaysAllowDepositsForRewardsUser) {
   // Act & Assert
-  for (int i = 0; i < static_cast<int>(mojom::AdType::kMaxValue); ++i) {
-    for (int j = 0; j < static_cast<int>(mojom::ConfirmationType::kMaxValue);
-         ++j) {
+  for (size_t i = 0; i < static_cast<size_t>(mojom::AdType::kMaxValue); ++i) {
+    for (size_t j = 0;
+         j < static_cast<size_t>(mojom::ConfirmationType::kMaxValue); ++j) {
       EXPECT_TRUE(IsAllowedToDeposit(static_cast<mojom::AdType>(i),
                                      static_cast<mojom::ConfirmationType>(j)));
     }
@@ -33,8 +35,8 @@ TEST_F(BraveAdsAccountUtilTest,
   test::DisableBraveRewards();
 
   // Act & Assert
-  for (int i = 0; i < static_cast<int>(mojom::ConfirmationType::kMaxValue);
-       ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::ConfirmationType::kMaxValue); ++i) {
     EXPECT_TRUE(IsAllowedToDeposit(mojom::AdType::kInlineContentAd,
                                    static_cast<mojom::ConfirmationType>(i)));
   }
@@ -46,8 +48,8 @@ TEST_F(BraveAdsAccountUtilTest,
   test::DisableBraveRewards();
 
   // Act & Assert
-  for (int i = 0; i < static_cast<int>(mojom::ConfirmationType::kMaxValue);
-       ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::ConfirmationType::kMaxValue); ++i) {
     EXPECT_TRUE(IsAllowedToDeposit(mojom::AdType::kPromotedContentAd,
                                    static_cast<mojom::ConfirmationType>(i)));
   }
@@ -59,8 +61,8 @@ TEST_F(BraveAdsAccountUtilTest,
   test::DisableBraveRewards();
 
   // Act & Assert
-  for (int i = 0; i < static_cast<int>(mojom::ConfirmationType::kMaxValue);
-       ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::ConfirmationType::kMaxValue); ++i) {
     EXPECT_FALSE(IsAllowedToDeposit(mojom::AdType::kNewTabPageAd,
                                     static_cast<mojom::ConfirmationType>(i)));
   }
@@ -76,8 +78,8 @@ TEST_F(
   test::DisableBraveRewards();
 
   // Act & Assert
-  for (int i = 0; i < static_cast<int>(mojom::ConfirmationType::kMaxValue);
-       ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::ConfirmationType::kMaxValue); ++i) {
     EXPECT_TRUE(IsAllowedToDeposit(mojom::AdType::kNewTabPageAd,
                                    static_cast<mojom::ConfirmationType>(i)));
   }
@@ -89,8 +91,8 @@ TEST_F(BraveAdsAccountUtilTest,
   test::DisableBraveRewards();
 
   // Act & Assert
-  for (int i = 0; i < static_cast<int>(mojom::ConfirmationType::kMaxValue);
-       ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::ConfirmationType::kMaxValue); ++i) {
     EXPECT_FALSE(IsAllowedToDeposit(mojom::AdType::kNotificationAd,
                                     static_cast<mojom::ConfirmationType>(i)));
   }
@@ -103,8 +105,8 @@ TEST_F(
   test::DisableBraveRewards();
 
   // Act & Assert
-  for (int i = 0; i < static_cast<int>(mojom::ConfirmationType::kMaxValue);
-       ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::ConfirmationType::kMaxValue); ++i) {
     EXPECT_FALSE(IsAllowedToDeposit(mojom::AdType::kSearchResultAd,
                                     static_cast<mojom::ConfirmationType>(i)));
   }
@@ -120,8 +122,8 @@ TEST_F(
   test::DisableBraveRewards();
 
   // Act & Assert
-  for (int i = 0; i < static_cast<int>(mojom::ConfirmationType::kMaxValue);
-       ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::ConfirmationType::kMaxValue); ++i) {
     const auto confirmation_type = static_cast<mojom::ConfirmationType>(i);
 
     const bool is_allowed_to_deposit =
