+++ components/brave_ads/core/internal/creatives/notification_ads/creative_notification_ad_test_util.cc
@@ -13,12 +13,12 @@
 
 namespace brave_ads::test {
 
-CreativeNotificationAdList BuildCreativeNotificationAds(int count) {
-  CHECK_GT(count, 0);
+CreativeNotificationAdList BuildCreativeNotificationAds(size_t count) {
+  CHECK_GT(count, 0U);
 
   CreativeNotificationAdList creative_ads;
 
-  for (int i = 0; i < count; ++i) {
+  for (size_t i = 0; i < count; ++i) {
     CreativeNotificationAdInfo creative_ad = BuildCreativeNotificationAd(
         /*should_generate_random_uuids=*/true);
     creative_ad.segment = kSegments[i % kSegments.size()];
