+++ components/brave_ads/core/internal/serving/prediction/model_based/sampling/creative_ad_model_based_predictor_sampling_unittest.cc
@@ -61,7 +61,7 @@ TEST_F(BraveAdsCreativeAdModelBasedPredictorSamplingTest,
   creative_ad_predictors.push_back(creative_ad_predictor_3);
 
   // Act & Assert
-  for (int i = 0; i < 10; ++i) {
+  for (size_t i = 0; i < 10; ++i) {
     EXPECT_EQ(creative_ad_predictor_2.creative_ad,
               MaybeSampleCreativeAd(creative_ad_predictors));
   }
@@ -91,7 +91,7 @@ TEST_F(BraveAdsCreativeAdModelBasedPredictorSamplingTest,
   int creative_ad_1_count = 0;
   int creative_ad_2_count = 0;
 
-  for (int i = 0; i < 25; ++i) {
+  for (size_t i = 0; i < 25; ++i) {
     // P(X>1) > 0.99999999 with X~Bin(n=25, p=0.5), i.e. less than 1 in 100M
     // tests are expected to fail.
     const std::optional<CreativeNotificationAdInfo> creative_ad =
