+++ components/brave_ads/core/internal/user_engagement/ad_events/ad_event_test_util.h
@@ -6,6 +6,7 @@
 #ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_USER_ENGAGEMENT_AD_EVENTS_AD_EVENT_TEST_UTIL_H_
 #define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_USER_ENGAGEMENT_AD_EVENTS_AD_EVENT_TEST_UTIL_H_
 
+#include <cstddef>
 #include <vector>
 
 #include "brave/components/brave_ads/core/mojom/brave_ads.mojom-forward.h"
@@ -25,7 +26,7 @@ void RecordAdEvents(
 
 void RecordAdEvents(const AdInfo& ad,
                     mojom::ConfirmationType mojom_confirmation_type,
-                    int count);
+                    size_t count);
 
 }  // namespace test
