+++ components/brave_ads/core/internal/user_engagement/ad_events/search_result_ads/search_result_ad_event_handler_util_for_rewards_unittest.cc
@@ -75,8 +75,9 @@ TEST_F(BraveAdsSearchResultAdEventHandlerUtilForRewardsTest,
           /*should_generate_random_uuids=*/true);
 
   // Act
-  for (int i = 0;
-       i < static_cast<int>(mojom::SearchResultAdEventType::kMaxValue); ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::SearchResultAdEventType::kMaxValue);
+       ++i) {
     const auto mojom_ad_event_type =
         static_cast<mojom::SearchResultAdEventType>(i);
     if (mojom_ad_event_type !=
@@ -124,8 +125,9 @@ TEST_F(BraveAdsSearchResultAdEventHandlerUtilForRewardsTest,
           /*should_generate_random_uuids=*/true);
 
   // Act & Assert
-  for (int i = 0;
-       i < static_cast<int>(mojom::SearchResultAdEventType::kMaxValue); ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::SearchResultAdEventType::kMaxValue);
+       ++i) {
     const auto mojom_ad_event_type =
         static_cast<mojom::SearchResultAdEventType>(i);
     EXPECT_TRUE(IsAllowedToFireAdEvent(mojom_creative_ad, mojom_ad_event_type));
@@ -140,8 +142,9 @@ TEST_F(BraveAdsSearchResultAdEventHandlerUtilForRewardsTest,
           /*should_generate_random_uuids=*/true);
 
   // Act & Assert
-  for (int i = 0;
-       i < static_cast<int>(mojom::SearchResultAdEventType::kMaxValue); ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::SearchResultAdEventType::kMaxValue);
+       ++i) {
     const auto mojom_ad_event_type =
         static_cast<mojom::SearchResultAdEventType>(i);
     EXPECT_TRUE(IsAllowedToFireAdEvent(mojom_creative_ad, mojom_ad_event_type));
